Вынес append всего контейнера из 25 фото за forEach для исключения ненужного дублирования
